canbus& driver: split frequency converter and magnetic async fron contrller
