+++ modules/canbus/vehicle/diamond/diamond_controller.h
@@ -37,32 +37,6 @@ namespace apollo {
 namespace canbus {
 namespace diamond {
 
-float getLatdev(int dec) {
-  long long int bin = 0;
-  int temp = dec;
-  long long j = 1;
-  while (temp) {
-    bin = bin + j * (temp % 2);
-    temp = temp / 2;
-    j = j * 10;
-  }
-  std::string s = std::to_string(bin);
-  while (s.size() < 16) {
-    s = '0' + s;
-  }
-  int sum_activated = 0;
-  int sum_id = 0;
-  for (int i = 0; i < 16; i++) {
-    if (s[i] == '1') {
-      sum_id += 16 - i;
-      sum_activated += 1;
-    }
-  }
-  auto lat_dev_mgs =
-      static_cast<float>(sum_id) / static_cast<float>(sum_activated) - 8.5;
-  return lat_dev_mgs;
-}
-
 class DiamondController final : public VehicleController {
  public:
   explicit DiamondController(){};
@@ -164,7 +138,7 @@ class DiamondController final : public VehicleController {
   std::thread thread_mangetic_;
   std::future<void> async_action_;
 
-  // 变频器 485通信 设备
+  // 变频器设备 485通信
   std::unique_ptr<Uart> steer_front = nullptr;
   std::unique_ptr<Uart> steer_rear = nullptr;
